requirements.txt
pdfminer.six==20240706
reportlab==4.2.5
pytesseract==0.3.13
Pillow==10.4.0
pdf2image==1.17.0